### 2026-01-16 #9786 @editorialbot
[PRE REVIEW]: ArviZ: a modular and flexible library for exploratory analysis of Bayesian models
**Submitting author:** @aloctavodia (<a href="http://orcid.org/0000-0001-7419-8978">Osvaldo Martin</a>)
**Repository:** https://github.com/arviz-devs/arviz
**Branch with paper.md** (empty if default branch): paper
**Version:** 1.0.0rc0
