### 2026-01-16 7b0344a
update requirements (#2514)
* update requirements

* update workflows

not sure they will succeed due to the figshare issues as those fixes
aren't included here, but figured it would be interesting to see either
way.
