### 2026-01-16 #9786 @crvernon
> ⚠️ [An error happened when generating the pdf](https://github.com/openjournals/joss-papers/actions/runs/21070055947?check_suite_focus=true). Problem with affiliations for Aki Vehtari, perhaps the affiliations index need quoting?.

@aloctavodia - please check out the above paper compilation error.  Let me know when it is fixed.  Thanks.
