### 2026-01-16 #9786 @crvernon
👋 @matt-graham  - are you able to take this one on as editor?

@aloctavodia will have to update the submission to the new JOSS requirements, but once this is done you can proceed with finding reviewers.
