### 2026-01-17 abe207e
add missing sections, fix missing DOIs
