+++ paper/references.bib
@@ -126,7 +126,8 @@ encoding = {UTF-8},
   title = {Implicitly adaptive importance sampling},
   journal = {Stat Comp},
   volume = 31,
-  number = 16
+  number = 16,
+  doi= {https://doi.org/10.1007/s11222-020-09982-2},
 }
 
 @article{Gelman_2019,
@@ -217,6 +218,7 @@ url = {https://plot.ly},
       archivePrefix={arXiv},
       primaryClass={cs.HC},
       url={https://arxiv.org/abs/2408.16702},
+      doi={https://doi.org/10.48550/arXiv.2408.16702},
 }
 
 @book{Martin_2024,
