### 2026-01-17 #9786 @arfon
> @openjournals/dev has this editor been given credentials yet?

Looks like the GitHub team permissions hadn't been loaded by Editorialbot...
