### 2026-01-17 #9786 @arfon
@matt-graham – you can assign yourself with `@editorialbot assign me as editor`
