### 2026-01-17 #9786 @matt-graham
> @matt-graham – you can assign yourself with `@editorialbot assign me as editor`

Thanks, was just looking up what the command was in documentation 😅!
