### 2026-01-17 #9786 @danielskatz
But an editorialbot command needs to be the first thing in a new comment...
