### 2026-01-19 #9786 @matt-graham
Hi @aloctavodia, thanks for updating the paper to fix the compilation error and adding the new sections for the updated scope. This all looks in line now so I will begin the process of finding reviewers.
