### 2026-01-19 #9786 @matt-graham
@markur4 @smutch @njtierney can you review this submission for JOSS?
