### 2026-01-19 c3761ee
update AI usage disclosure and move to the bottom
