### 2026-01-20 #9786 @smutch
@matt-graham I'd be happy to review. Just to double check for transparency though - I have one very minor bug-fix commit in the repo to be reviewed. I just want to make sure that's ok? I am not associated or affiliated with PyMC or any of the maintainers though.
