### 2026-01-21 #9786 @OriolAbril
To add the full context, the paper is for ArviZ which is only a dependency of PyMC. ArviZ is an independent project from PyMC with each having their own governing bodies and governance rules.

That being said, both are libraries for Bayesian modeling, so there is some overlap between the teams, with PyMC for example, 3 of the 7 people authoring this paper are also authors on the relatively recent [PeerJ PyMC paper](https://peerj.com/articles/cs-1516/). At the same time, two other authors of this paper are Stan developers.
