### 2026-01-21 #9786 @matt-graham
> I'd be happy to review. Just to double check for transparency though - I have one very minor bug-fix commit in the repo to be reviewed. I just want to make sure that's ok? I am not associated or affiliated with PyMC or any of the maintainers though.

Hi @smutch - thanks for being willing to review and for your transparency in declaring your prior involvement in the project.  I have checked with the wider editorial team and we do not think that contributing a bug-fix would constitute a conflict of interest under [the JOSS policy](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html#joss-conflict-of-interest-policy) so we would be very glad to have you as a reviewer. I still need to find one or two additional reviewers for this submission, but I will assign you as a reviewer for now and you will receive a notification when the review starts in a separate issue.
