### 2026-01-22 #9786 @matt-graham
@gchure @vankesteren can you review this submission for JOSS?
