### 2026-01-22 #9786 @gchure
Hi @matt-graham, I've used arviz for quite some time, so I would love to review this work if a 3-4 week turnaround is palatable (a bit pressed by other obligations at the moment).
