### 2026-01-22 #9786 @matt-graham
Hi @gchure. Thanks for being willing to review this submission! Yes 3-4 week turnaround would be fine - we generally request JOSS reviews are completed with 4-6 weeks.

I will wait a little bit longer to see if anyone else I tagged is also able to review this before starting the review.
