### 2026-01-23 #9786 @gmarupilla
@matt-graham, I'm interested in reviewing this submission. I have experience working with Bayesian modeling workflows and model diagnostics in Python. Happy to review.
