### 2026-01-23 #9786 @vankesteren
I'd be happy to review! Unlike @gchure I actually haven't used Arviz extensively (other than reading the landing page once or twice) so I think we'll be complementary 👍
