### 2026-01-23 #9786 @matt-graham
Hi @GMarupilla and @vankesteren. Thank you both very much for offering to review this submission! I am now in the fortunate position of having too many people available to review as `gchure` and `smutch` have also already agreed to review, and we generally aim for two or three reviewers per submission.

I can assign you both as reviewers for this submission if that is your preference. Alternatively I am currently looking for one additional reviewer for another submission #9704 which has a similar Bayesian inference and modelling application area, so if either of you would be able to review that submission instead that would be great.
