### 2026-01-23 #9786 @gmarupilla
Thanks for letting me know. I’m okay with that and happy to review https://github.com/openjournals/joss-reviews/issues/9704
