### 2026-01-23 #9889 @editorialbot
[REVIEW]: ArviZ: a modular and flexible library for exploratory analysis of Bayesian models
**Submitting author:** @aloctavodia (<a href="http://orcid.org/0000-0001-7419-8978">Osvaldo Martin</a>)
**Repository:** https://github.com/arviz-devs/arviz
**Branch with paper.md** (empty if default branch): paper
**Version:** v1.0.0
