### 2026-01-23 f3b14ba
remove prefix from DOIs
