+++ paper/references.bib
@@ -1,6 +1,6 @@
 @article{Kumar_2019,
 doi = {10.21105/joss.01143},
-url = {https://doi.org/10.21105/joss.01143},
+url = {10.21105/joss.01143},
 year = {2019}, publisher = {The Open Journal},
 volume = {4},
 number = {33},
@@ -127,7 +127,7 @@ encoding = {UTF-8},
   journal = {Stat Comp},
   volume = 31,
   number = 16,
-  doi= {https://doi.org/10.1007/s11222-020-09982-2},
+  doi= {10.1007/s11222-020-09982-2},
 }
 
 @article{Gelman_2019,
@@ -176,7 +176,7 @@ year  = {2019}
   year      = {2017},
   publisher = {Ubiquity Press},
   doi       = {10.5334/jors.148},
-  url       = {https://doi.org/10.5334/jors.148}
+  url       = {10.5334/jors.148}
 }
 
 @article{Hunter_2007,
@@ -218,7 +218,7 @@ url = {https://plot.ly},
       archivePrefix={arXiv},
       primaryClass={cs.HC},
       url={https://arxiv.org/abs/2408.16702},
-      doi={https://doi.org/10.48550/arXiv.2408.16702},
+      doi={10.48550/arXiv.2408.16702},
 }
 
 @book{Martin_2024,
@@ -233,7 +233,7 @@ url = {https://plot.ly},
 }
 
 
-@article{emcee, doi = {10.21105/joss.01864}, url = {https://doi.org/10.21105/joss.01864}, year = {2019}, publisher = {The Open Journal}, volume = {4}, number = {43}, pages = {1864}, author = {Daniel Foreman-Mackey and Will M. Farr and Manodeep Sinha and Anne M. Archibald and David W. Hogg and Jeremy S. Sanders and Joe Zuntz and Peter K. g. Williams and Andrew R. j. Nelson and Miguel de Val-Borro and Tobias Erhardt and Ilya Pashchenko and Oriol Abril Pla}, title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC}, journal = {Journal of Open Source Software} }
+@article{emcee, doi = {10.21105/joss.01864}, url = {10.21105/joss.01864}, year = {2019}, publisher = {The Open Journal}, volume = {4}, number = {43}, pages = {1864}, author = {Daniel Foreman-Mackey and Will M. Farr and Manodeep Sinha and Anne M. Archibald and David W. Hogg and Jeremy S. Sanders and Joe Zuntz and Peter K. g. Williams and Andrew R. j. Nelson and Miguel de Val-Borro and Tobias Erhardt and Ilya Pashchenko and Oriol Abril Pla}, title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC}, journal = {Journal of Open Source Software} }
 
 
 @book{eabm_2025,
@@ -244,7 +244,7 @@ url = {https://plot.ly},
   publisher    = {Zenodo},
   version      = {v0.3.0},
   doi          = {10.5281/zenodo.15127548},
-  url          = {https://doi.org/10.5281/zenodo.15127548},
+  url          = {10.5281/zenodo.15127548},
                   },
 
 
