### 2026-01-23 #9889 @aloctavodia
Thanks @smutch, @gchure, and @vankesteren for taking the time to review our paper.

This is already mentioned in the README, but to avoid any confusion, I want to clarify that the repository https://github.com/arviz-devs/arviz is for the arviz-metapackage. This was also the repository for the legacy version of ArviZ. It is fine if you open issues there. But if you inspect the main branch, you will notice that it is almost empty, and most of the code lives in these 3 repositories.

* https://github.com/arviz-devs/arviz-base
* https://github.com/arviz-devs/arviz-stats
* https://github.com/arviz-devs/arviz-plots
