### 2026-01-24 8c7ce78
Apply suggestions from code review
Co-authored-by: Seth Axen <seth@sethaxen.com>
