+++ paper/references.bib
@@ -6,7 +6,7 @@ volume = {4},
 number = {33},
 pages = {1143},
 author = {Ravin Kumar and Colin Carroll and Ari Hartikainen and Osvaldo Martin},
-title = {ArviZ a unified library for exploratory analysis of Bayesian models in Python},
+title = {{ArviZ} a unified library for exploratory analysis of {Bayesian} models in {Python}},
 journal = {Journal of Open Source Software}
 }
 
@@ -36,7 +36,7 @@ journal = {Journal of Open Source Software}
 }
 
 @article{Phan_2019,
-  title={Composable Effects for Flexible and Accelerated Probabilistic Programming in NumPyro},
+  title={Composable Effects for Flexible and Accelerated Probabilistic Programming in {NumPyro}},
   author={Phan, Du and Pradhan, Neeraj and Jankowiak, Martin},
   journal={arXiv preprint arXiv:1912.11554},
   year={2019}
@@ -72,7 +72,7 @@ journal = {Journal of Open Source Software}
 }
 
 @article{Sailynoja_2025,
-  title={Recommendations for visual predictive checks in Bayesian workflow},
+  title={Recommendations for visual predictive checks in {Bayesian} workflow},
   author={S{\"a}ilynoja, Teemu and Johnson, Andrew R and Martin, Osvaldo A and Vehtari, Aki},
   journal={arXiv:2503.01509},
   doi={10.48550/arXiv.2503.01509},
@@ -158,7 +158,7 @@ year  = {2019}
 }
 
 @article{Capretto_2022,
- title={Bambi: A Simple Interface for Fitting Bayesian Linear Models in Python},
+ title={Bambi: A Simple Interface for Fitting {Bayesian} Linear Models in {Python}},
  volume={103},
  number={15},
  journal={Journal of Statistical Software},
@@ -168,7 +168,7 @@ year  = {2019}
 }
 
 @article{Hoyer_2017,
-  title     = {xarray: {N-D} labeled arrays and datasets in {Python}},
+  title     = {{xarray}: {N-D} labeled arrays and datasets in {Python}},
   author    = {Hoyer, S. and J. Hamman},
   journal   = {Journal of Open Research Software},
   volume    = {5},
@@ -181,7 +181,7 @@ year  = {2019}
 
 @article{Hunter_2007,
   Author    = {Hunter, J. D.},
-  Title     = {Matplotlib: A 2D graphics environment},
+  Title     = {Matplotlib: A {2D} graphics environment},
   Journal   = {Computing in Science \& Engineering},
   Volume    = {9},
   Number    = {3},
@@ -195,14 +195,14 @@ year  = {2019}
 }
 
 @manual{Bokeh_2018,
-title = {Bokeh: Python library for interactive visualization},
+title = {Bokeh: {Python} library for interactive visualization},
 author = {{Bokeh Development Team}},
 year = {2018},
 url = {https://bokeh.pydata.org/en/latest/},
 }
 
 @online{plotly_2015,
-author = {Plotly Technologies Inc.},
+author = {{Plotly Technologies Inc.}},
 title = {Collaborative data science},
 publisher = {Plotly Technologies Inc.},
 address = {Montreal, QC},
@@ -211,7 +211,7 @@ url = {https://plot.ly},
 }
 
 @misc{Guo_2024,
-      title={VMC: A Grammar for Visualizing Statistical Model Checks},
+      title={{VMC}: A Grammar for Visualizing Statistical Model Checks},
       author={Ziyang Guo and Alex Kale and Matthew Kay and Jessica Hullman},
       year={2024},
       eprint={2408.16702},
@@ -238,7 +238,7 @@ url = {https://plot.ly},
 
[…]
