### 2026-01-24 91ef530
doc: Update setup guidelines for pyproject.toml (#2517)
