### 2026-01-24 4e27959
update stan reference
