+++ paper/references.bib
@@ -28,11 +28,15 @@ journal = {Journal of Open Source Software}
 }
 
 @article{stan,
-  title = {Stan: {{A Probabilistic Programming Language}} | {{Carpenter}} | {{Journal}} of {{Statistical Software}}},
-	shorttitle = {{}},
-  doi = {10.18637/jss.v076.i01},
-  language = {en-US},
-  keywords = {Bayesian inference,algorithmic differentiation,probabilistic programming,Stan},
+ title={Stan: A Probabilistic Programming Language},
+ volume={76},
+ url={https://www.jstatsoft.org/index.php/jss/article/view/v076i01},
+ doi={10.18637/jss.v076.i01},
+ number={1},
+ journal={Journal of Statistical Software},
+ author={Carpenter, Bob and Gelman, Andrew and Hoffman, Matthew D. and Lee, Daniel and Goodrich, Ben and Betancourt, Michael and Brubaker, Marcus and Guo, Jiqiang and Li, Peter and Riddell, Allen},
+ year={2017},
+ pages={1-32}
 }
 
 @article{Phan_2019,
