+++ paper/references.bib
@@ -180,7 +180,7 @@ year  = {2019}
   year      = {2017},
   publisher = {Ubiquity Press},
   doi       = {10.5334/jors.148},
-  url       = {10.5334/jors.148}
+  url       = {https://doi.org/10.5334/jors.148}
 }
 
 @article{Hunter_2007,
@@ -237,7 +237,17 @@ url = {https://plot.ly},
 }
 
 
-@article{emcee, doi = {10.21105/joss.01864}, url = {10.21105/joss.01864}, year = {2019}, publisher = {The Open Journal}, volume = {4}, number = {43}, pages = {1864}, author = {Daniel Foreman-Mackey and Will M. Farr and Manodeep Sinha and Anne M. Archibald and David W. Hogg and Jeremy S. Sanders and Joe Zuntz and Peter K. g. Williams and Andrew R. j. Nelson and Miguel de Val-Borro and Tobias Erhardt and Ilya Pashchenko and Oriol Abril Pla}, title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC}, journal = {Journal of Open Source Software} }
+@article{emcee,
+         doi = {10.21105/joss.01864},
+         url = {https://doi.org/10.21105/joss.01864},
+         year = {2019},
+         publisher = {The Open Journal},
+         volume = {4},
+         number = {43},
+         pages = {1864},
+         author = {Daniel Foreman-Mackey and Will M. Farr and Manodeep Sinha and Anne M. Archibald and David W. Hogg and Jeremy S. Sanders and Joe Zuntz and Peter K. g. Williams and Andrew R. j. Nelson and Miguel de Val-Borro and Tobias Erhardt and Ilya Pashchenko and Oriol Abril Pla},
+         title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC},
+         journal = {Journal of Open Source Software} }
 
 
 @book{eabm_2025,
@@ -248,7 +258,7 @@ url = {https://plot.ly},
   publisher    = {Zenodo},
   version      = {v0.3.0},
   doi          = {10.5281/zenodo.15127548},
-  url          = {10.5281/zenodo.15127548},
+  url          = {https://doi.org/10.5281/zenodo.15127548},
                   },
 
 @software{arvizjl_2025,
