### 2026-01-24 b0cdf46
update documentation generation guidelines
