### 2026-01-27 #9889 @vankesteren
@aloctavodia when reviewing / trying out the software for the purposes of this paper, I expect that I need to use

```py
import arviz.preview as az
```

for now, and the preview version would become the main version upon publication of the paper? Is this how you're planning it? Following https://github.com/arviz-devs/arviz/issues/2491

I guess my question is: are we reviewing ArviZ 1.0?
