### 2026-01-27 #9889 @aloctavodia
Hi @vankesteren

Currently, you can install arviz 0.23.1 (and a few older versions too) and get access to the "new arviz" using the `preview` submodel like `import arviz.preview as az`. This assumes you have also installed arviz-base, arviz-stats, and arviz-plots (or some subset).

We also have a 1.0.rc0 version, you can install it with `pip install "arviz==1.0.rc0"`. If you do this, this should install all the arviz_* subpackages. And by doing `import arviz as az`, you will get access to the all the new functions.
