### 2026-01-27 #9889 @vankesteren
## Paper comments checklist

After going through the current version of the paper, here are my main comments as a checklist. Feel free to respond to the items here; I will go over the paper again when all the items have been dealt with / responded to! Then I will update my review checklist.

### summary section

- [ ] For new users: the summary needs a sentence on what arviz does beyond "exploratory analyses of Bayesian models". There are many people who don't exactly know what this means. So some information about: it does model diagnostics, inferential plots, model comparison, and predictive plots (or so) would help here.

### statement of need section

- [ ] The second paragraph of the statement of need focuses a lot on technical aspects and less on the actual needs/improvements/impact. For example in the statement of need. In the second sentence of the second paragraph, I expect the need for the redesign to be stated. "while ArviZ has enabled ..., recent developments in probabilistic programming (and available data structures?? xarray) such as... led to the need for an overhaul of the software". Also, on line 29 "Additionally, the new design facilitates the use of components as modular building blocks for custom analyses" does not tell me much. What does this enable that ArviZ did not enable before? A reader does not need a lot of detail here, but the context is important.

### research impact section

- [ ] Fix weird in-text reference to Stan on line 37

- [ ] What is the function of the paragraph starting on line 41? Could be removed or rewritten into the research impact statement paragraph, something like "the maturity of ArviZ has also led to other initiatives such as ArviZ.jl, PreliZ, and educational resources." Please don't refer to "additional packages that are still in an experimental phase": this does not add to the paper.

### state of the field section

- [ ] there is no "state of the field" section. In it, I would expect a reference to the `posterior` package (in R/Stan community) which does similar things to your data structure, as well as the `bayesplot` package which does similar things to your visualizations. I think is a missed opportunity. Some comparing / contrasting would be nice. (And by this I don't mean: we do x and y which these do not; just to give context to the reader who might be familiar with these packages)

### software design section
- [ ] line 49 "with improved design as described next": this reads a bit strange, and also this is not described immediately after? Please fix this.

- [ ] I really like the move to a more general data structure (xarray datatree) but the second paragraph of the software design section (line 50) could describe why you did this in one sentence: maintainability, interoperability, simplicity of the ArviZ library, better serialization support...

- [ ] line 58-61: if possible, it would be nice to include a very simple example of each in-line so we can compare it, e.g., to compute the mean of a parameter's posterior samples. Feel free to ignore as this is also later in the examples section

### examples section
- [ ] example 1 uses array_stats but I think in arviz.preview it's been renamed (to stats?) or the functionality has been incorporated somehow? I suggest doing import arviz as az on line 81, and then on line 85 do az.ess(samples, chain_axis=0, draw_axis=1)

- [ ] example on line 95: I am not getting the same plot output with filled bars, but rather a lineplot when I rerun this code:
