### 2026-01-27 78c638b
Update pages under Contribution types section
