+++ docs/source/contributing/contributing_prs.md
@@ -1,10 +1,9 @@
 # Contributing code via pull requests
 
 While issue reporting is valuable, we strongly encourage users who are
-inclined to do so to submit patches for new or existing issues via pull
-requests. This is particularly the case for simple fixes, such as typos
-or tweaks to documentation, which do not require a heavy investment
-of time and attention.
+inclined to do so to submit patches for new or existing issues via pull requests.
+This is particularly the case for simple fixes, such as typos or tweaks to documentation,
+which do not require a heavy investment of time and attention.
 
 Contributors are also encouraged to contribute new code to enhance ArviZ's
 functionality, also via pull requests.
@@ -13,10 +12,22 @@ to ensure that any new contribution does not strongly overlap with existing
 functionality and open a "Feature Request" issue before starting to work
 on the new feature.
 
+:::{note}
+Since version 1.0.0, ArviZ is organized as a metapackage that brings together functionality
+from three subpackages: [`arviz-base`](https://github.com/arviz-devs/arviz-base),
+[`arviz-stats`](https://github.com/arviz-devs/arviz-stats), and
+[`arviz-plots`](https://github.com/arviz-devs/arviz-plots).
+
+As a result, contributing to ArviZ typically means contributing to one of these subpackages rather
+than to the main ArviZ repository itself.
+When browsing issues or pull requests, keep in mind that most activity happens in the
+subpackage repositories.
+:::
+
 (steps_before_working)=
 
 ## Steps before starting work
-Before starting work on a pull request double-check that no one else
+Before starting work on a pull request, double-check that no one else
 is working on the ticket in both issue tickets and pull requests.
 
 ArviZ is a community-driven project and always has multiple people working
@@ -25,45 +36,47 @@ that we all make the most of our time and we don't have two contributors
 working on the same changes. Let's see what to do when you encounter the following scenarios:
 
 ### If an issue ticket exists
-If an issue exists check the ticket to ensure no one else has started working on it. If you are first to start work, comment on the ticket to make it evident to others. If the comment looks old or abandoned leave a comment asking if you may start work.
+If an issue exists, check the ticket to ensure no one else has started working on it.
+If you are first to start work, comment on the ticket to make it evident to others.
+If the comment looks old or abandoned, leave a comment asking if you may start work.
 
 ### If an issue ticket doesn't exist
-Open an issue ticket for the issue and state that you'll be solving the issue with a pull request. Optionally create a pull request and add `[WIP]` in the title to indicate Work in Progress.
+Open an issue ticket for the issue and state that you'll be solving the issue with a pull request.
+Optionally create a pull request and add `[WIP]` in the title to indicate Work in Progress.
 
 ### In the event of a conflict
-In the event of two or more people working on the same issue, the general precedence will go to the person who first commented on the issue. If no comments it will go to the first person to submit a PR for review. Each situation will differ though, and the core contributors will make the best judgment call if needed.
+In the event of two or more people working on the same issue,
+the general precedence will go to the person who first commented on the issue.
+If there are no comments, it will go to the first person to submit a PR for review.
+Each situation will differ though, and the core contributors will make the best judgment call if needed.
 
 ### If the issue ticket has someone assigned to it
-If the issue is assigned then precedence goes to the assignee. However, if there has been no activity for 2 weeks from the assignment date, the ticket is open for all again and can be unassigned.
+If the issue is assigned, then precedence goes to the assignee.
+However, if there has been no activity for 2 weeks from the assignment date,
+the ticket is open for all again and can be unassigned.
 
 (dev_summary)=
 ## Development process - summary
 
[…]
