### 2026-01-27 f69333f
pr tutorial improvements
