+++ tox.ini
@@ -6,6 +6,8 @@ envlist =
 # See https://tox.readthedocs.io/en/latest/example/package.html#flit
 isolated_build = True
 isolated_build_env = build
+labels =
+    dev = check, cleandocs, docs, viewdocs, test-namespace
 
 [gh-actions]
 python =
@@ -14,6 +16,7 @@ python =
     3.14: test-namespace
 
 [testenv]
+description = Run ArviZ metapackage tests
 basepython =
     # See https://github.com/tox-dev/tox/issues/1548
     {check,docs,cleandocs,viewdocs,test-namespace}: python3
@@ -28,7 +31,7 @@ commands =
     pytest {env:PYTEST_MARKERS:} {env:PYTEST_EXTRA_ARGS:} {posargs:-vv}
 
 [testenv:check]
-description = perform style checks
+description = Perform style checks
 extras =
     check
 allowlist_externals =
@@ -38,7 +41,7 @@ commands =
     python -m build
 
 [testenv:docs]
-description = build HTML docs
+description = Build HTML docs
 setenv =
     READTHEDOCS_PROJECT = arviz
     READTHEDOCS_VERSION = local
@@ -49,14 +52,14 @@ commands =
     sphinx-build -d "{toxworkdir}/docs_doctree" docs/source "{toxworkdir}/docs_out" --color -v -bhtml
 
 [testenv:cleandocs]
-description = clean HTML outputs docs
+description = Clean HTML doc build outputs
 skip_install = true
 allowlist_externals = rm
 commands =
     rm -r "{toxworkdir}/docs_out" "{toxworkdir}/docs_doctree" "{toxworkdir}/jupyter_execute" docs/source/_miniatures
 
 [testenv:viewdocs]
-description = open HTML docs
+description = View HTML docs with the default browser
 skip_install = true
 commands =
     python -m webbrowser "{toxworkdir}/docs_out/index.html"
