### 2026-01-27 0ce72ae
add h5py as doc dependency
