### 2026-01-27 f7baca8
try dev arviz-plots as plot_pair fix
