+++ pyproject.toml
@@ -27,7 +27,7 @@ dynamic = ["version", "description"]
 dependencies = [
   "arviz_base >=0.7.0",
   "arviz_stats[xarray] >=0.7.0",
-  "arviz_plots >=0.7.0",
+  "arviz-plots @ git+https://github.com/arviz-devs/arviz-plots",
 ]
 
 [tool.flit.module]
