### 2026-01-27 f14376e
Use system cache directory for timestamp file
