### 2026-01-27 f69174c
Make timestamp file write atomic
