### 2026-01-27 1b5f8bc
fixup! Make timestamp file write atomic
