### 2026-01-27 8ab27ff
update doc building instruccions
