+++ tox.ini
@@ -51,6 +51,16 @@ commands =
     python docs/scripts/homepage_miniatures.py
     sphinx-build -d "{toxworkdir}/docs_doctree" docs/source "{toxworkdir}/docs_out" --color -v -bhtml
 
+[testenv:livedocs]
+description = Build live preview of HTML docs with automatic updates
+base = docs
+extras =
+    doc
+    sphinx-autobuild
+commands =
+    python docs/scripts/homepage_miniatures.py
+    sphinx-autobuild -d "{toxworkdir}/docs_doctree" docs/source "{toxworkdir}/docs_out" --color -v -bhtml
+
 [testenv:cleandocs]
 description = Clean HTML doc build outputs
 skip_install = true
