### 2026-01-27 3def7ae
Change version from 1.0.0 to 1.0
