### 2026-01-27 3036a2a
Use 'package' instead of 'subpackage' when refering to arviz-* packages.
