+++ docs/source/contributing/contributing_prs.md
@@ -14,14 +14,14 @@ on the new feature.
 
 :::{note}
 Since version 1.0, ArviZ is organized as a metapackage that brings together functionality
-from three subpackages: [`arviz-base`](https://github.com/arviz-devs/arviz-base),
+from three packages: [`arviz-base`](https://github.com/arviz-devs/arviz-base),
 [`arviz-stats`](https://github.com/arviz-devs/arviz-stats), and
 [`arviz-plots`](https://github.com/arviz-devs/arviz-plots).
 
-As a result, contributing to ArviZ typically means contributing to one of these subpackages rather
+As a result, contributing to ArviZ typically means contributing to one of these packages rather
 than to the main ArviZ repository itself.
 When browsing issues or pull requests, keep in mind that most activity happens in the
-subpackage repositories.
+corresponding package repositories.
 :::
 
 (steps_before_working)=
