+++ docs/source/contributing/issue_reports.md
@@ -3,7 +3,7 @@
 
 We appreciate being notified of problems with the existing ArviZ code.
 We prefer that issues be filed on the
-Github Issue Tracker of the corresponding sub-package repository
+Github Issue Tracker of the corresponding package repository
 ([`arviz-base`](https://github.com/arviz-devs/arviz-base/issues),
 [`arviz-stats`](https://github.com/arviz-devs/arviz-stats/issues),
 [`arviz-plots`](https://github.com/arviz-devs/arviz-stats/issues))
