### 2026-01-27 ffddf78
Remove architecture.md and developing_in_docker.md
