### 2026-01-27 7a9b204
Simplify pull request checklist
