+++ docs/source/contributing/docstrings.md
@@ -2,8 +2,9 @@
 # Docstring style
 
 Docstrings should follow the
-[numpy docstring guide](https://numpydoc.readthedocs.io/en/latest/format.html). Read more about it {ref}`docstring_formatting`.
-In ArviZ docs, docstrings consit of five main sections, i.e.,
+[numpy docstring guide](https://numpydoc.readthedocs.io/en/latest/format.html).
+
+In ArviZ docs, docstrings consist of five main sections, i.e.,
 1. Short summary
 2. Parameters
 3. Returns
@@ -17,19 +18,27 @@ Docstrings follow the same guide for adding references as the other docs.
 For adding references to external libraries functions and objects, see {ref}`reference_external_libs`. For referencing ArviZ objects, follow {ref}`reference_arviz_objects`.
 
 ## See Also
-In ArviZ docs, we have a lot of interconnected functions both within the library and with external libraries and it can take a lot of time to search for the related functions. It is crucial to add the See Also section to save users time.
-For adding the _See Also_ docstring section, you just need to add the function name. Sphinx will
-automatically add links to other ArviZ objects and functions listed in the _See Also_
+In ArviZ docs, we have a lot of interconnected functions both within the library and with external
+libraries and it can take a lot of time to search for the related functions.
+It is crucial to add the See Also section to save users time.
+For adding the _See Also_ docstring section, you just need to add the function name.
+Sphinx will automatically add links to other ArviZ objects and functions listed in the _See Also_
 section.
 
-For example, let's add {func}`~arviz.hdi` and {func}`~arviz.plot_ppc` in the _See Also_ section.
+For example, let's add {func}`~arviz_stats.hdi` and {func}`~arviz_plots.plot_ppc` in the _See Also_ section.
 
 ```
     See Also
     --------
-    hdi : Calculate highest density interval (HDI) of array for given probability.
-    plot_ppc : plot for posterior/prior predictive checks.
+    arviz_stats.hdi : Calculate highest density interval (HDI) of array for given probability.
+    arviz_plots.plot_ppc : plot for posterior/prior predictive checks.
 ```
 
 ## Kwargs parameters
-All the kwargs parameters in {ref}`plots <plot_api>` modules are passed to the matplotlib or bokeh functions. While writing their description, the functions to which they are being passed must be mentioned. In order to check or add those functions, the process is the same for all the kwargs arguments. Let's read the step-by-step guide for `backend_kwargs` as an example, [here](https://github.com/arviz-devs/arviz/wiki/ArviZ-Hacktoberfest-2021).
+
+All the kwargs parameters in {ref}`arviz_plots:api` modules are passed to the
+Matplotlib or Bbokeh functions. While writing their description, the functions to which they are
+being passed must be mentioned. In order to check or add those functions, the process is the same
+for all the kwargs arguments.
+Let's read the step-by-step guide for `backend_kwargs` as an example,
+[here](https://github.com/arviz-devs/arviz/wiki/ArviZ-Hacktoberfest-2021).
