### 2026-01-27 f76dc7d
Update in-depth explanations
