### 2026-01-27 2ef314d
Fix references and add h5py to doc environment
