### 2026-01-28 #9889 @matt-graham
@aloctavodia - generally for JOSS we recommend authors address any issues raised by reviewers as they are raised rather than waiting for all reviews to come in a batch, so if you are able to start addressing @vankesteren's comments whenever you have a chance to that would be great.
