### 2026-01-28 #9889 @vankesteren
### Authorship clarification

dear @aloctavodia and team. During my review I noticed something in the authorship that I would like to see clarified. As context, here is the official [JOSS policy on authorship](https://joss.readthedocs.io/en/latest/submitting.html#authorship):

> Purely financial (such as being named on an award) and organizational (such as general supervision of a research group) contributions are not considered sufficient for co-authorship of JOSS submissions, but active project direction and other forms of non-code contributions are. The authors themselves assume responsibility for deciding who should be credited with co-authorship, and co-authors must always agree to be listed.

The point I want to raise is that while @avehtari is listed as the senior/last author on the paper, his _visible contribution_ to the software is very small: a [single documentation commit](https://github.com/arviz-devs/arviz/commit/ab82f80f7862058fc3d1430d4a972f247bf16e78), which is smaller than other contributors who have not been included as co-authors.

I realise that per JOSS policy, we let authors themselves decide who should be eligible for co-authorship, but I nonetheless would like some confirmation that prof. Vehtari's inclusion as an author is warranted following this policy (for example, pointing to potential non-visible / non-code contributions). I am aware of his position in the field and I think his contributions to the Bayesian workflow makes his work relevant to the present paper, but this alone is not enough for co-authorship on this particular work.
