### 2026-01-28 #9889 @aloctavodia
Thanks @matt-graham, we will start addressing the reviewers' comments.

@vankesteren Regarding authorship, Aki Vehtari is considered a core contributor to ArviZ even though he has not contributed code directly. His main contributions have been through technical advice, conceptual guidance, and discussions on implementation details and best practices. Additionally, he has helped maintain active communication with closely related R projects, including loo, posterior, bayesplot, and priorsense.
