+++ paper/references.bib
@@ -284,3 +284,41 @@ url = {https://joss.theoj.org/papers/10.21105/joss.05499},
 volume = {8},
 year = {2023}
 }
+
+@Misc{bayesplot0,
+  title = {bayesplot: Plotting for Bayesian Models},
+  author = {Jonah Gabry and Tristan Mahr},
+  year = {2025},
+  note = {R package version 1.15.0},
+  url = {https://mc-stan.org/bayesplot/},
+}
+
+@Article{bayesplot1,
+  title = {Visualization in Bayesian workflow},
+  author = {Jonah Gabry and Daniel Simpson and Aki Vehtari and Michael Betancourt and Andrew Gelman},
+  year = {2019},
+  journal = {J. R. Stat. Soc. A},
+  volume = {182},
+  issue = {2},
+  pages = {389-402},
+  doi = {10.1111/rssa.12378},
+}
+
+@book{gelman_2013,
+    address = {Boca Raton},
+    edition = {3 edition},
+    title = {Bayesian {Data} {Analysis}, {Third} {Edition}},
+    isbn = {978-1-4398-4095-5},
+    publisher = {Chapman and Hall/CRC},
+    author = {Gelman, Andrew and Carlin, John B. and Stern, Hal S. and Dunson, David B. and Vehtari, Aki and Rubin, Donald B.},
+    month = nov,
+    year = {2013},
+}
+
+@Misc{loo,
+  title = {loo: Efficient leave-one-out cross-validation and WAIC for Bayesian models},
+  year = {2025},
+  note = {R package version 2.9.0},
+  url = {https://mc-stan.org/loo/},
+}
+
