+++ paper/references.bib
@@ -322,3 +322,12 @@ year = {2023}
   url = {https://mc-stan.org/loo/},
 }
 
+@article{Padilla_2021,
+  title={Uncertainty Visualization},
+  author={Padilla, Lace and Kay, Matthew and Hullman, Jessica},
+  journal={Wiley StatsRef: Statistics Reference Online},
+  year={2021},
+  publisher={Wiley Online Library},
+  doi={10.1002/9781118445112.stat08296},
+  url={http://space.ucmerced.edu/Downloads/publications/Uncertainty_Visualization_Padilla_Kay_Hullman_2022.pdf}
+}
