### 2026-02-03 b818746
update example to use qds, add missing references, small tweaks
