+++ paper/paper.md
@@ -46,25 +46,27 @@ bibliography: references.bib
 
 # Summary
 
-When working with Bayesian models, a range of related tasks must be addressed beyond inference itself. These include tasks such as visualizing the model results along with their uncertainty, diagnosing the quality of MCMC samples, model criticism or model comparison. We collectively refer to these activities as exploratory analysis of Bayesian models.
+When working with Bayesian models, a range of related tasks must be addressed beyond inference itself. These include diagnosing the quality of MCMC samples, model criticism, model comparison, etc. We collectively refer to these activities as exploratory analysis of Bayesian models.
 
 In this work, we present a redesigned version of `ArviZ`, a Python package for exploratory analysis of Bayesian models. The redesign emphasizes greater user control and modularity. This redesign delivers a more flexible and efficient toolkit for exploratory analysis of Bayesian models. With its renewed focus on modularity and usability, `ArviZ` is well-positioned to remain an essential tool for Bayesian modelers in both research and applied settings.
+
 
 # Statement of need
 
 Probabilistic programming has emerged as a powerful paradigm for statistical modeling, accompanied by a growing ecosystem of tools for model specification and inference. Effective modeling requires robust support for uncertainty visualization, sampling diagnostics, model comparison, and model checking [@Gelman_2020; @Martin_2024; @Guo_2024]. `ArviZ` addresses this gap by providing a unified, backend-agnostic library to perform these tasks. The original `ArviZ` paper [@Kumar_2019] described the landscape of probabilistic programming tools at the time and the need for a unified, backend-agnostic library for exploratory analysis - a need that has only grown as the ecosystem has expanded.
 
-The methods implemented in `ArviZ` are grounded in well-established statistical principles and provide robust, interpretable diagnostics and visualizations [@Vehtari_2017; @Gelman_2019; @Dimitriadis_2021; @Paananen_2021; @Padilla_2021; @Vehtari_2021; @Sailynoja_2022; @Kallioinen_2023; @Sailynoja_2025]. The redesigned version furthers these goals by keeping the easy-to-use interface for regular users and introducing more powerful tooling for power users and developers of Bayesian tools. These updates align with recent developments in the probabilistic programming field. Additionally, the new design facilitates the use of components as modular building blocks for custom analyses. This frequent user request was difficult to accommodate under the old framework.
+The methods implemented in `ArviZ` are grounded in well-established statistical principles and provide robust, interpretable diagnostics and visualizations [@Vehtari_2017; @Gelman_2019; @Dimitriadis_2021; @Paananen_2021; @Padilla_2021; @Vehtari_2021; @Sailynoja_2022; @Kallioinen_2023; @Sailynoja_2025]. As modern Bayesian practice has increasingly emphasized iterative, simulation-based workflows for model building, checking, validation, and comparison, the original ArviZ design became harder to adapt to these emerging needs. The redesigned version furthers ArviZ’s goals by introducing an easier-to-use interface for routine analyses alongside more powerful, flexible tooling for advanced users and developers of Bayesian software. These updates align ArviZ with the current demands of Bayesian modeling and workflow development
+
 
 # State of the field
 
-In the Python Bayesian ecosystem, ArviZ occupies a niche comparable to tools in the R/Stan community such as posterior [@gelman_2013;@Vehtari_2021], loo [@Vehtari_2017;@loo], bayesplot [@bayesplot0;@bayesplot1], and priorsense [@Kallioinen_2023], sharing similar goals while reflecting different language ecosystems and workflows.
[…]
