+++ paper/references.bib
@@ -331,3 +331,82 @@ year = {2023}
   doi={10.1002/9781118445112.stat08296},
   url={http://space.ucmerced.edu/Downloads/publications/Uncertainty_Visualization_Padilla_Kay_Hullman_2022.pdf}
 }
+
+
+@inproceedings{kay_2016,
+author = {Kay, Matthew and Kola, Tara and Hullman, Jessica R. and Munson, Sean A.},
+title = {When (ish) is My Bus? User-centered Visualizations of Uncertainty in Everyday, Mobile Predictive Systems},
+year = {2016},
+isbn = {9781450333627},
+publisher = {Association for Computing Machinery},
+address = {New York, NY, USA},
+url = {https://doi.org/10.1145/2858036.2858558},
+doi = {10.1145/2858036.2858558},
+pages = {5092–5103},
+numpages = {12},
+keywords = {uncertainty visualization, transit predictions, mobile interfac-es, end-user visualization, dotplots},
+location = {San Jose, California, USA},
+series = {CHI '16}
+}
+
+@article{kay_2024,
+  author = {Matthew Kay},
+  title = {{ggdist}: Visualizations of Distributions and Uncertainty in the Grammar of Graphics},
+  journal = {IEEE Transactions on Visualization and Computer Graphics},
+  year = {2024},
+  volume = {30},
+  number = {1},
+  pages = {414--424},
+  doi = {10.1109/TVCG.2023.3327195},
+}
+
+
+@article{harris_2020,
+ title         = {Array programming with {NumPy}},
+ author        = {Charles R. Harris and K. Jarrod Millman and St{\'{e}}fan J.
+                 van der Walt and Ralf Gommers and Pauli Virtanen and David
+                 Cournapeau and Eric Wieser and Julian Taylor and Sebastian
+                 Berg and Nathaniel J. Smith and Robert Kern and Matti Picus
+                 and Stephan Hoyer and Marten H. van Kerkwijk and Matthew
+                 Brett and Allan Haldane and Jaime Fern{\'{a}}ndez del
+                 R{\'{i}}o and Mark Wiebe and Pearu Peterson and Pierre
+                 G{\'{e}}rard-Marchant and Kevin Sheppard and Tyler Reddy and
+                 Warren Weckesser and Hameer Abbasi and Christoph Gohlke and
+                 Travis E. Oliphant},
+ year          = {2020},
+ month         = sep,
+ journal       = {Nature},
+ volume        = {585},
+ number        = {7825},
+ pages         = {357--362},
+ doi           = {10.1038/s41586-020-2649-2},
+ publisher     = {Springer Science and Business Media {LLC}},
+ url           = {https://doi.org/10.1038/s41586-020-2649-2}
+}
+
+
+@ARTICLE{virtanen_2020,
+  author  = {Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E. and
+            Haberland, Matt and Reddy, Tyler and Cournapeau, David and
+            Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and
+            Bright, Jonathan and {van der Walt}, St{\'e}fan J. and
+            Brett, Matthew and Wilson, Joshua and Millman, K. Jarrod and
+            Mayorov, Nikolay and Nelson, Andrew R. J. and Jones, Eric and
+            Kern, Robert and Larson, Eric and Carey, C J and
+            Polat, {\.I}lhan and Feng, Yu and Moore, Eric W. and
+            {VanderPlas}, Jake and Laxalde, Denis and Perktold, Josef and
+            Cimrman, Robert and Henriksen, Ian and Quintero, E. A. and
[…]
