### 2026-02-03 bd7cfeb
Apply suggestions from code review
Co-authored-by: Tomás Capretto <tomicapretto@gmail.com>
