### 2026-02-03 4115b8a
go over the rest of how-to pages
