### 2026-02-04 254d502
Use system cache directory for timestamp file (#2523)
* Use system cache directory for timestamp file

* Make timestamp file write atomic

* fixup! Make timestamp file write atomic

* use ndownloader.figshare.com for all links

* add platformdirs to requirements

---------

Co-authored-by: Oriol (ProDesk) <oriol.abril.pla@gmail.com>
