+++ requirements.txt
@@ -9,3 +9,4 @@ h5netcdf>=1.0.2
 h5py
 typing_extensions>=4.1.0
 xarray-einstats>=0.3
+platformdirs
