### 2026-02-04 bb0872a
update version to be able to release properly
