### 2026-02-04 e85410b
Update docs/source/contributing/contributing_prs.md
Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>
