### 2026-02-04 0e3195a
Update docs/source/contributing/docstrings.md
Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>
