### 2026-02-04 17eca88
Update docs/source/contributing/contributing_prs.md
Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>
