### 2026-02-04 a824796
Update docs/source/contributing/diataxis_for_arviz.md
Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>
