### 2026-02-04 a948237
work on aliases and redirects for docs and update migration guide (#2509)
* extend objects.inv with aliases

* ensure version is either stable or latest

* add redirects for api pages

* add redirects for InferenceData API pages

* some updates to migration guide

* add idata->dt focused migration advice

* include pyproject updates to fix doc build

* address review comments
