### 2026-02-04 6470f96
Update Contribution types, Reference, and In depth explanations in Contributing guide (#2521)
* Update pages under Contribution types section

* Change version from 1.0.0 to 1.0

* Use 'package' instead of 'subpackage' when refering to arviz-* packages.

* Typo

* Unify TODO

* Remove architecture.md and developing_in_docker.md

* Simplify pull request checklist

* Update pages in 'Reference'

* Update in-depth explanations

* Fix references and add h5py to doc environment

* Restore miniatures to branch base

* Update docs/source/contributing/contributing_prs.md

Co-authored-by: Oriol Abril-P
[…]
