+++ docs/source/contributing/pr_checklist.md
@@ -5,64 +5,45 @@ We recommend that your contribution complies with the following guidelines befor
 
 * If your pull request addresses an issue, please use the pull request title to describe the issue and mention the issue number in the pull request description. This will make sure a link back to the original issue is created.
 
-* All public methods must have informative docstrings with sample usage when appropriate.
+* Please use a Draft Pull Request to indicate an incomplete contribution or work in progress.
+Draft PRs may be useful to (1) signal you are working on something and avoid duplicated work,
+(2) request early feedback on functionality or API, or (3) seek collaborators.
 
-* Please prefix the title of incomplete contributions with `[WIP]` (to indicate a work in progress). WIPs may be useful to (1) indicate you are working on something to avoid duplicated work, (2) request a broad review of functionality or API, or (3) seek collaborators.
+* Run any of the pre-existing examples notebooks that contain analyses that would be affected by
+your changes to ensure that nothing breaks. This is a useful opportunity to not only check your
+work for bugs that might not be revealed by unit test, but also to show how your contribution
+improves ArviZ for end users.
 
-* All other tests pass when everything is rebuilt from scratch.
-See {ref}`developing_in_docker` for information on running the test suite locally.
+* All public functions and methods must have informative NumPy-style docstrings,
+including Examples and, when appropriate, See Also, References, and Notes.
+You can refer to {func}`~arviz_plots.plot_dist` and {func}`~arviz_plots.plot_energy` for good examples.
 
-* When adding additional plotting functionality, provide at least one example script in the [arviz/examples/](https://github.com/arviz-devs/arviz/tree/main/examples) folder. Have a look at other examples for reference. Examples should demonstrate why the new functionality is useful in practice and, if possible, compare it to other methods available in ArviZ.
+* New functionality must be covered by tests, and tests context must follow the
+[pytest fixture pattern](https://docs.pytest.org/en/latest/fixture.html#fixture).
 
-* Added tests follow the [pytest fixture pattern](https://docs.pytest.org/en/latest/fixture.html#fixture).
+* Your code follows the style guidelines of the project:
 
-* Documentation and high-coverage tests are necessary for enhancements to be accepted.
-
-* Documentation follows Numpy style guide.
-
-* Run any of the pre-existing examples in ``docs/source/notebooks`` that contain analyses that would be affected by your changes to ensure that nothing breaks. This is a useful opportunity to not only check your work for bugs that might not be revealed by unit test, but also to show how your contribution improves ArviZ for end users.
-
-* If modifying a plot, render your plot to inspect for changes and copy image in the pull request message on Github.
-
-You can also check for common programming errors with the following
-tools:
-
-* Save plots as part of tests. Plots will be saved to a directory named `test_images` by default.
-
-  ```bash
-  $ pytest arviz/tests/base_tests/<name of test>.py --save
+  ```shell
+  tox -e check
   ```
 
-* Optionally save plots to a user named directory. This is useful for comparing changes across branches.
+* All **tests must pass**.
 
-  ```bash
-  $ pytest arviz/tests/base_tests/<name of test>.py --save user_defined_directory
+  ```shell
+  tox -e py312
   ```
 
+  You can replace `py312` with the version that matches your environment (e.g., `py311`, `py313`, etc.).
 
-* Code coverage **cannot** decrease. Coverage can be checked with **pytest-cov** package:
+  Moreover, given that libraries have their own testing tasks, it's recommended to check their specific
+  testing guidelines:
 
-  ```bash
[…]
