### 2026-02-04 c9a63bb
Update Tutorials and How-to contributing pages (#2522)
* pr tutorial improvements

* add h5py as doc dependency

* try dev arviz-plots as plot_pair fix

* update doc building instruccions

* Apply suggestions from code review

Co-authored-by: Tomás Capretto <tomicapretto@gmail.com>

* go over the rest of how-to pages

---------

Co-authored-by: Tomás Capretto <tomicapretto@gmail.com>
