+++ docs/source/contributing/pr_tutorial.md
@@ -5,13 +5,38 @@ The preferred workflow for contributing to ArviZ is to fork
 the [GitHub repository](https://github.com/arviz-devs/arviz/),
 clone it to your local machine, and develop on a feature branch.
 
+This tutorial applies to any of these repositories:
+
+* [arviz-devs/arviz](https://github.com/arviz-devs/arviz)
+* [arviz-devs/arviz-base](https://github.com/arviz-devs/arviz-base)
+* [arviz-devs/arviz-stats](https://github.com/arviz-devs/arviz-stats)
+* [arviz-devs/arviz-plots](https://github.com/arviz-devs/arviz-plots)
+
+
 (pr_steps)=
 ## Steps
 
-1. Fork the [project repository](https://github.com/arviz-devs/arviz/) by clicking on the 'Fork' button near the top right of the main repository page. This creates a copy of the code under your GitHub user account.
+:::{admonition} Have you already gone through this tutorial?
+:class: dropdown tip
+
+If this is not your first time going over this tutorial and you want to
+work on your 2nd+ PR you can skip steps 1-3, and instead do:
+
+```bash
+git checkout main
+git fetch upstream
+git rebase upstream/main
+```
+
+This will make sure you are on the `main` branch of your local repository and sync it
+with the one on GitHub. After this you can continue with {ref}`step 4 <feature_branch_step_pr>`
+to create your feature branch from the up-to-date `main`
+:::
+
+1. Fork the project repository by clicking on the 'Fork' button near the top right of the main repository page. This creates a copy of the code under your GitHub user account.
 
 (fork_step_pr)=
-2. Clone your fork of the ArviZ repo from your GitHub account to your local disk.
+2. Clone your fork of the target ArviZ repository from your GitHub account to your local disk.
 
    ::::{tab-set}
 
@@ -19,7 +44,7 @@ clone it to your local machine, and develop on a feature branch.
    :sync: ssh
 
    ```
-   git clone git@github.com:<your GitHub handle>/arviz.git
+   git clone git@github.com:<your GitHub handle>/<arviz repo>.git
    ```
    :::
 
@@ -27,7 +52,7 @@ clone it to your local machine, and develop on a feature branch.
    :sync: https
 
    ```
-   git clone https://github.com/<your GitHub handle>/arviz.git
+   git clone https://github.com/<your GitHub handle>/<arviz repo>.git
    ```
    :::
 
@@ -41,8 +66,8 @@ clone it to your local machine, and develop on a feature branch.
    :sync: ssh
 
    ```
-   cd arviz
-   git remote add upstream git@github.com:arviz-devs/arviz.git
+   cd <arviz repo>
+   git remote add upstream git@github.com:arviz-devs/<arviz repo>.git
    ```
    :::
 
@@ -50,8 +75,8 @@ clone it to your local machine, and develop on a feature branch.
    :sync: https
 
    ```
-   cd arviz
-   git remote add upstream https://github.com/arviz-devs/arviz.git
+   cd <arviz repo>
+   git remote add upstream https://github.com/arviz-devs/<arviz repo>.git
    ```
    :::
 
@@ -66,28 +91,59 @@ clone it to your local machine, and develop on a feature branch.
 
    ```{warning}
    Always create a new ``feature`` branch before making any changes. Make your changes
-   in the ``feature`` branch. It's good practice to never routinely work on the ``main`` branch of any repository.
+   in the ``feature`` branch. It's good practice to never routinely work on the ``main`` branch of any repository
+   and we have a pre-commit check that prevents committing to `main`
    ```
 
[…]
