+++ docs/source/contributing/running_benchmarks.md
@@ -1,9 +1,0 @@
-# Running the benchmark suite
-
-To run the **benchmark tests** do the following:
-
-```bash
-$ pip install asv
-$ cd arviz
-$ asv run
-```
