+++ pyproject.toml
@@ -50,7 +50,6 @@ plotly = ["arviz_plots[plotly]"]
 
 # dev related blocks
 doc = [
-    "h5netcdf[h5py]",
     "h5netcdf",
     "h5py",
     "jupyter-sphinx",
@@ -63,6 +62,7 @@ doc = [
     "sphinx-design",
     "sphinx-notfound-page",
     "sphinxcontrib-youtube",
+    "sphinx-togglebutton",
     "sphobjinv",
 ]
 test = ["pytest"]

[…]
