### 2026-02-09 cf25b22
Split migration guide into Migration Guide + What's New in 1.0 (#2527)
- Add short migration_guide.md (need-to-know checklist)
- Add whats_new_1_0.md (full narrative and examples)
- Update user guide index toctree to include both docs

Co-authored-by: Cursor <cursoragent@cursor.com>
