+++ docs/source/user_guide/whats_new_1_0.md
@@ -1,0 +1,642 @@
+---
+jupyter:
+  jupytext:
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.3'
+      jupytext_version: 1.19.1
+  kernelspec:
+    display_name: Python 3 (ipykernel)
+    language: python
+    name: python3
+---
+
+(whats_new_1_0)=
+# What's new in ArviZ 1.0?
+
+For a quick upgrade checklist of breaking changes and minimal code edits, see the {ref}`migration_guide`.
+
+Below is the full picture of what changed and what's new in 1.0.
+
+```python editable=true slideshow={"slide_type": ""} tags=["remove-cell"]
+import xarray as xr
+xr.set_options(display_expand_attrs=False, display_expand_coords=False);
+```
+
+```python editable=true slideshow={"slide_type": ""}
+import arviz as az
+```
+
+Check all 3 libraries have been exposed correctly:
+
+```python
+print(az.info)
+```
+
+## ArviZ 1.0: modularity and optional dependencies
+
+We have been working on refactoring ArviZ to allow more flexibility and extensibility of its elements
+while keeping as much as possible a friendly user-interface that gives sensible results with little to no arguments.
+
+One important change is enhanced modularity. Everything will still be available through a common namespace `arviz`,
+but ArviZ will now be composed of 3 smaller libraries:
+
+* [arviz-base](https://arviz-base.readthedocs.io/en/latest/) data related functionality, including converters from different PPLs.
+* [arviz-stats](https://arviz-stats.readthedocs.io/en/latest/) for statistical functions and diagnostics.
+* [arviz-plots](https://arviz-plots.readthedocs.io/en/latest/) for visual checks built on top of arviz-stats and arviz-base.
+
+Each library has a minimal set of dependencies, with a lot of functionality built on top of optional dependencies.
+This keeps ArviZ smaller and easier to install as you can install only the components you really need. The main examples are:
+
+* `arviz-base` has no I/O library as a dependency, but you can use `netcdf4`, `h5netcdf` or `zarr` to read and write your data, allowing you to install only the one you need.
+* `arviz-plots` has no plotting library as a dependency, but it can generate plots with `matplotlib`, `bokeh` or `plotly` if they are installed.
+
+## `arviz-base`
+
+<!-- #region -->
+### Credible intervals and rcParams
+
+Some global configuration settings have changed. For example, the default credible interval probability (`ci_prob`) has been updated from 0.94 to 0.89. Using 0.89 produces intervals with lower variability, leading to more stable summaries. At the same time, keeping a non-standard value (rather than 0.90 or 0.95) serves as a friendly reminder that the choice of interval can depend on the problem at hand.
+
+In addition, a new setting `ci_kind` has been introduced, which defaults to "eti" (equal-tailed interval). This controls the method used to compute credible intervals. The alternative is "hdi" (highest density interval), which was previously the default.
+
+
+Defaults set via `rcParams` are not fixed rules, they're meant to be adjusted to fit the needs of your analysis. `rcParams` offers a convenient way to establish global defaults for your workflow, while most functions that compute credible intervals also provide `ci_prob` and `ci_kind` arguments to override these settings locally.
+
+
+You can check all default settings with:
+<!-- #endregion -->
+
+```python
+az.rcParams
+```
+
+### `DataTree`
+One of the main differences is that the `arviz.InferenceData` object doesn't exist anymore.
+`arviz-base` uses {class}`xarray.DataTree` instead. This is a new data structure in xarray,
+so it might still have some rough edges, but it is much more flexible and powerful.
[…]
