+++ docs/source/user_guide/migration_guide.md
@@ -1,0 +1,230 @@
+---
+jupyter:
+  jupytext:
+    text_representation:
+      extension: .md
+      format_name: markdown
+      format_version: '1.3'
+      jupytext_version: 1.19.1
+  kernelspec:
+    display_name: Python 3 (ipykernel)
+    language: python
+    name: python3
+---
+
+(migration_guide)=
+# ArviZ migration guide
+
+ArviZ 1.0 introduces breaking changes: a new data structure (DataTree), new defaults, and some API renames. This page is a short checklist of what you need to migrate. For the full story—why things changed, new features, and examples—see {ref}`whats_new_1_0`.
+
+ArviZ is now split into three libraries (`arviz-base`, `arviz-stats`, `arviz-plots`). Unless you know you only need one part, you can get all three by importing the main package as usual:
+
+```python
+import arviz as az
+```
+
+Optional: confirm all three libraries are exposed:
+
+```python
+print(az.info)
+```
+
+
+### Credible intervals and rcParams
+
+1. **Interval type (`ci_kind`):** In previous versions, summaries and interval-based functions used the highest-density interval (HDI) by default. In 1.0 the default is the equal-tailed interval (ETI), which tends to give more stable summaries. 
+
+Functions that accept `ci_kind` include {func}`~arviz.summary`, {func}`~arviz.bayesian_r2`, {func}`~arviz.ci_in_rope`, {func}`~arviz.loo_r2`, {func}`~arviz.residual_r2`, {func}`~arviz_plots.plot_convergence_dist`, {func}`~arviz_plots.plot_dist`, {func}`~arviz_plots.plot_forest`, {func}`~arviz_plots.plot_lm`, {func}`~arviz_plots.plot_ppc_interval`, {func}`~arviz_plots.plot_ppc_tstat`, and {func}`~arviz_plots.plot_psense_dist`.
+
+   **What to do:** If you're fine with ETI, you don't need to change anything (results may differ). If you want HDI, add `ci_kind="hdi"` to your calls or set `az.rcParams["stats.ci_kind"] = "hdi"` once.
+
+
+2. **Parameter name (`hdi_prob` → `ci_prob`):** The argument for credible-interval probability was renamed from `hdi_prob` to `ci_prob`, because "ci" (credible interval) is the general term for either HDI or ETI. In 1.0 the old name is **removed** (it raises an error). 
+
+Functions that accept `ci_prob` include: {func}`~arviz.summary`, {func}`~arviz.bayesian_r2`, {func}`~arviz.ci_in_rope`, {func}`~arviz.loo_r2`, {func}`~arviz.residual_r2`, {func}`~arviz_plots.plot_convergence_dist`, {func}`~arviz_plots.plot_dist`, {func}`~arviz_plots.plot_lm`, {func}`~arviz_plots.plot_ppc_pava`, {func}`~arviz_plots.plot_ppc_pava_residuals`, {func}`~arviz_plots.plot_ppc_rootogram`, {func}`~arviz_plots.plot_ppc_tstat`, {func}`~arviz_plots.plot_psense_dist`.
+
+   **What to do:** In functions that use the keyword argument `hdi_prob`, replace it with `ci_prob`. If you use an `rcParam` that includes `hdi_prob`, as in `az.rcParams["stats.hdi_prob"]`, replace it with `ci_prob`. 
+   
+   
+3. **Default value (0.94 → 0.89):** The default interval probability is now 0.89 instead of 0.94, for more stable summaries.
+
[…]
