+++ tox.ini
@@ -66,7 +66,7 @@ description = Clean HTML doc build outputs
 skip_install = true
 allowlist_externals = rm
 commands =
-    rm -r "{toxworkdir}/docs_out" "{toxworkdir}/docs_doctree" "{toxworkdir}/jupyter_execute" docs/source/_miniatures
+    rm -r "{toxworkdir}/docs_out" "{toxworkdir}/docs_doctree" "{toxworkdir}/jupyter_execute" docs/source/_miniatures docs/source/savefig
 
 [testenv:viewdocs]
 description = View HTML docs with the default browser
