### 2026-02-10 34005d2
Address PR #2528 review: reorder, rcParams, MyST, headers, URLs, examples
- Reorder migration guide by impact (plots first, I/O last)
- Shorten Dimensions section; consolidate rcParams into one section at end
- Fix header hierarchy (## for sections); use MyST format for both docs
- Update issue.md link to migration_guide.md; document URL decisions in plan
- Add before/after examples (I/O, plot kwargs); explicit ic_compare_method in compare example

Co-authored-by: Cursor <cursoragent@cursor.com>
