+++ docs/source/user_guide/whats_new_1_0.md
@@ -3,7 +3,7 @@ jupyter:
   jupytext:
     text_representation:
       extension: .md
-      format_name: markdown
+      format_name: myst
       format_version: '1.3'
       jupytext_version: 1.19.1
   kernelspec:
