+++ docs/source/user_guide/migration_guide.md
@@ -3,7 +3,7 @@ jupyter:
   jupytext:
     text_representation:
       extension: .md
-      format_name: markdown
+      format_name: myst
       format_version: '1.3'
       jupytext_version: 1.19.1
   kernelspec:
@@ -23,54 +23,57 @@ ArviZ is now split into three libraries (`arviz-base`, `arviz-stats`, `arviz-plo
 import arviz as az
 ```
 
-Optional: confirm all three libraries are exposed:
+## Plot names and signatures
 
-```python
-print(az.info)
-```
+Some plot functions were renamed or split; others kept the same name but have updated arguments and defaults. Use the following as a quick reference; see {ref}`whats_new_1_0` for the full inventory and new/removed functions.
 
+**Renamed or restructured:**
 
-### Credible intervals and rcParams
+| ArviZ &lt;1 | ArviZ ≥1 |
+|-------------|----------|
+| `plot_bpv` | `plot_ppc_pit`, `plot_ppc_tstat` |
+| `plot_dist_comparison` | `plot_prior_posterior` |
+| `plot_ecdf` | `plot_dist`, `plot_ecdf_pit` |
+| `plot_ess` | `plot_ess`, `plot_ess_evolution` |
+| `plot_forest` | `plot_forest`, `plot_ridge` |
+| `plot_ppc` | `plot_ppc_dist` |
+| `plot_posterior`, `plot_density` | `plot_dist` |
+| `plot_trace` | `plot_trace_dist`, `plot_trace_rank` |
 
-1. **Interval type (`ci_kind`):** In previous versions, summaries and interval-based functions used the highest-density interval (HDI) by default. In 1.0 the default is the equal-tailed interval (ETI), which tends to give more stable summaries. 
+**Same name, updated implementation/arguments:** `plot_autocorr`, `plot_bf`, `plot_compare`, `plot_energy`, `plot_khat`, `plot_lm`, `plot_loo_pit`, `plot_mcse`, `plot_pair`, `plot_parallel`, `plot_rank`.
 
-Functions that accept `ci_kind` include {func}`~arviz.summary`, {func}`~arviz.bayesian_r2`, {func}`~arviz.ci_in_rope`, {func}`~arviz.loo_r2`, {func}`~arviz.residual_r2`, {func}`~arviz_plots.plot_convergence_dist`, {func}`~arviz_plots.plot_dist`, {func}`~arviz_plots.plot_forest`, {func}`~arviz_plots.plot_lm`, {func}`~arviz_plots.plot_ppc_interval`, {func}`~arviz_plots.plot_ppc_tstat`, and {func}`~arviz_plots.plot_psense_dist`.
+**Removed:** Legacy `plot_dist`, `plot_kde` (functionality is in the new `plot_dist`), `plot_violin`. Some functions (e.g. `plot_elpd`, `plot_ppc_residuals`, `plot_ts`) are planned but not yet available.
 
-   **What to do:** If you're fine with ETI, you don't need to change anything (results may differ). If you want HDI, add `ci_kind="hdi"` to your calls or set `az.rcParams["stats.ci_kind"] = "hdi"` once.
+## Plot return type and kwargs
 
+1. **Return type:** Plot functions now return a {class}`~arviz_plots.PlotCollection` (or similar) instead of raw axes or figures. That gives consistent layout, faceting, and control over how plots are combined.
 
-2. **Parameter name (`hdi_prob` → `ci_prob`):** The argument for credible-interval probability was renamed from `hdi_prob` to `ci_prob`, because "ci" (credible interval) is the general term for either HDI or ETI. In 1.0 the old name is **removed** (it raises an error). 
+   **What's affected:** All `plot_*` functions. Code that used the return value (e.g. to get axes or save a figure).
 
[…]
