### 2026-02-10 d4cb22e
Stop tracking local planning docs (issue.md, plan.md)
Co-authored-by: Cursor <cursoragent@cursor.com>
