+++ docs/source/user_guide/plan.md
@@ -1,154 +1,0 @@
-# Plan: Split migration guide into Migration Guide + What's New in 1.0
-
-## Steps
-
-- [x] **1. Create GitHub issue**  
-  Propose splitting the current migration guide into two documents: a short "need to know" migration guide and a user-focused "What's new in ArviZ 1.0?" document. Link to this plan and the outline below.
-
-- [x] **2. Create PR (after issue is agreed)**  
-  Use **jupytext** for the split and revisions so we keep editable notebooks in the repo while doing the heavy editing in markdown:
-
-  1. **Convert existing doc to markdown:** Use jupytext to convert `migration_guide.ipynb` to `.md` (e.g. `migration_guide.md`).
-  2. **Split the markdown:** Create two markdown files from the converted content—one for the **migration guide** (need-to-know only) and one for **"What's new in ArviZ 1.0?"** (informative)—following the outline below.
-  3. **Revise both docs:** Edit the two `.md` files (trim migration guide, expand/organize What's new, add cross-links).
-  4. **Convert back to notebooks:** Use jupytext to convert both revised `.md` files to `.ipynb` for the built docs (e.g. `migration_guide.ipynb` and `whats_new_1_0.ipynb`).
-  5. **Finalize:** Cross-link the two docs in the text, update `index.rst` / nav so both are discoverable, and remove or archive the temporary `.md` files if desired (or keep them as jupytext pair if the project prefers).
-
-- [ ] **3. (Optional)**  
-  Add a short "For developers" or "Technical details" section or doc if any content is still too implementation-focused for the main user docs.
-
----
-
-## Address PR #2528 review (OriolAbril)
-
-Steps to address colleague feedback on the migration guide and doc format:
-
-- [x] **4. Reorder migration guide sections by expected impact**  
-  Change section order so the highest-impact breaking changes come first:
-  1. **Plot names and signatures** (1st — renames affect many users)
-  2. **Plot return type and kwargs** (return type + visuals/stats/pc_kwargs)
-  3. **DataTree (replaces InferenceData)** (order of 2 vs 3 can be debated)
-  4. **Credible intervals and rcParams** and/or **Model comparison (WAIC, compare default)** (ci_prob/ci_kind, WAIC removal, stacking)
-  5. **I/O (netcdf/zarr)** last — behavior stays similar; main impact is installing dependencies.
-
-- [x] **5. Revisit dim / sample_dims in the migration guide**  
-  Colleague notes that dim/sample_dims is "basically all new functionality"; in 0.x only `hdi` (via kwarg with xarray.apply_ufunc) and a couple of plots with `skip_dims` could reduce over something other than chain+draw. Consider **removing or greatly shortening** the dim/sample_dims entry on this short, breaking-changes-only page (keep full treatment in What's new).
-
-- [x] **6. Consolidate rcParams into one section (last)**  
-  Merge the scattered rcParams-related bullets into a **single general "rcParams and defaults" section** at the end of the migration guide. Include:
-  - Ways to control behavior: **user-level or project templates**, **`rcParams` at the start of your code**, or **per-call arguments**.
-  - Clarify that **we do not consider changing rcParam defaults a breaking change**; users who care strongly about defaults should use a user-level template.
-  - Note: the `hdi_prob` → `ci_prob` rename was done in preparation (existing arvizrc templates already use `ci_prob`); the default weighting method was changed to stacking a while ago, so that is unchanged. Link to 0.x defaults: [arvizrc.template (v0.x)](https://github.com/arviz-devs/arviz/blob/v0.x/arvizrc.template).
-
-- [x] **7. Use MyST format for both docs**  
[…]
