### 2026-02-10 b4afdac
fix redirects (#2529)
