### 2026-02-10 2673d0b
Fix typos (#2531)
Found via `codespell -S *.ipynb,CHANGELOG.md`
