### 2026-02-11 9e3b594
docs: fix typo "Bbokeh" → "Bokeh" in docstring guide (#2532)
Co-authored-by: droppingbeans <beanbot@example.com>
