### 2026-02-13 #9889 @smutch
With the recent updates to the manuscript based on @vankesteren's feedback, I think it passes all of the checklist requirements. It's well written and concise. I was able to reproduce the examples with the pre 1.0 release without any issues, installing `arviz[preview]` and importing using `import arviz.preview as az`.

I had a good play with the new plotting API to explore one of my own models and ran in to no issues. It was great to see that the plots don't suffer from the pre 1.0 minor bugbear of overlapping labels/titles 😀. Some of the additional plots (e.g. `plot_prior_posterior`) are great quality-of-live additions for the exploratory workflow as well.

I really like the `PlotCollection` API in particular. After consulting the documentation it was relatively straight forward to, for example, alter the `neutral_element` color in example:

```python
pc.get_viz("dist", "mu").set_color("k")
pc.get_viz("dist", "tau").set_color("k")
pc.get_viz("legend", "school").get_lines()[0].set_color("k")
```

I was, however, a little surprised that altering `pc.aes.color["neutral_color"]` didn't actually have any effect on the plot. It might be worth noting that changes to `pc.aes` are not reflected in the final plot?

Another minor point I noticed is that `PlotCollection` doesn't render automatically in a Marimo notebook without calling `show()`. The output is then in the stdout section below the code (instead of above as is the default in Marimo). This could be addressed by the addition of a `__display__` method ([see the Marimo docs here](https://docs.marimo.io/guides/integrating_with_marimo/displaying_objects/)). I'll also open an issue for this, but its purely a quality-of-life suggestion.

Most of my evaluation of `arviz-stats` was through the plotting library, however, the code is well structured and has extensive tests. The documentation also does a good job of explaining the benefit of splitting this off into a separate package with the array and xarray interfaces.

Great work that I'm sure will be well appreciated and utilised by the community. 👍
