### 2026-02-13 #9889 @smutch
https://github.com/arviz-devs/arviz-plots/issues/427
