### 2026-02-13 #9889 @OriolAbril
Thanks for the review @smutch! I am very glad you like `PlotCollection`. It has made things easier on the development side but that doesn't always match users too.

> I was, however, a little surprised that altering pc.aes.color["neutral_color"] didn't actually have any effect on the plot. It might be worth noting that changes to pc.aes are not reflected in the final plot?

We will clarify this on the docs. Plotting is eager, the plotting backends are called whenever the `.map` method is used, which is what batteries included functions do. So whatever is in `.aes` at that time affects the plots, changing it after the fact has no effect unless you continue adding more elements to the figure. It would have an effect if the attribute contents are modified beforehand, but that is harder to do with `plot_xyz` functions. You'd have to create a `PlotCollection` manually, modify the `aes` attribute and then pass this as `plot_xyz(..., plot_collection=created_object)`.

> After consulting the documentation it was relatively straight forward to, for example, alter the neutral_element color in example:

As you have seen in the docs, the goal of storing all visuals/artists/glyphs in the PlotCollection class is allowing modifications like this one. That being said, in this particular example I would pass extra arguments to the PlotCollection created by `plot_dist` to manually define the color cycle used in that figure. That is because here we aren't really modifying things punctually but instead requesting a different dimension -> color mapping. It would look like:

```python
pc = az.plot_dist(
    dt,
    kind="dot",
    visuals={"point_estimate_text": False, "dist":{"marker": "C6"}},
    aes={"color": ["school"]},
    color=["black"]+[f"C{i+1}" for i in range(8)]
)
```

---

Also thanks for the marimo issue, we all have looking into it on our todo list but have very little experience with it for now. It is very helpful to know about low hanging fruit improvements like this one.
