### 2026-02-13 #9889 @matt-graham
Thanks @smutch for completing your review checklist and the additional feedback on the submission, and to @OriolAbril for the thorough responses to @smutch's comments / questions.

@vankesteren I think some of the items that are currently unchecked on your review checklist relate to [your feedback in this comment](https://github.com/openjournals/joss-reviews/issues/9889#issuecomment-3804707908) which I believe @aloctavodia [has responded to in this comment](https://github.com/openjournals/joss-reviews/issues/9889#issuecomment-3845392112). Let us know if you think the changes made / responses address your points, and if there is anything still needing resolving.

@gchure thanks for starting on your review. I remember you mentioned in the pre-review that you had other engagements in the short term. Do you have a firmer idea now of when you will be able to complete your review by?
