+++ .pre-commit-config.yaml
@@ -23,3 +23,15 @@ repos:
   rev: 0.4.1
   hooks:
     - id: no-print-statements
+- repo: https://github.com/Lucas-C/pre-commit-hooks
+  rev: v1.5.5
+  hooks:
+    - id: insert-license
+      files: ^(src|tests)/.*\.py$
+      args:
+        - --license-filepath
+        - tools/APACHE_HEADER.txt
+        - --comment-style
+        - "#"
+        - --detect-license-in-X-top-lines=30
+        - --no-extra-eol
