### 2026-02-23 2cfcd5a
Define acronyms on first usage
