### 2026-02-23 1d34670
Use ssentence case for section headers
