### 2026-02-23 d7aaa29
Consistently use non-code formatted project name
