### 2026-02-23 d3aba30
Expand abbreviated journal names
