+++ paper/references.bib
@@ -58,7 +58,7 @@ journal = {Journal of Open Source Software}
                Paul Horsfall and
                Noah D. Goodman},
   title     = {Pyro: Deep Universal Probabilistic Programming},
-  journal   = {J. Mach. Learn. Res.},
+  journal   = {Journal of Machine Learning Research},
   volume    = {20},
   pages     = {28:1--28:6},
   year      = {2019},
@@ -86,7 +86,7 @@ journal = {Journal of Open Source Software}
 @article{Vehtari_2017,
   title={Practical {Bayesian} model evaluation using leave-one-out cross-validation and {WAIC}},
   author={Vehtari, Aki and Gelman, Andrew and Gabry, Jonah},
-  journal={Stat Comp},
+  journal={Statistics and Computing},
   doi={10.1007/s11222-016-9696-4},
   volume={27},
   pages={1413--1432},
@@ -107,7 +107,7 @@ journal = {Journal of Open Source Software}
 	title = {Graphical test for discrete uniformity and its applications in goodness-of-fit evaluation and multiple sample comparison},
 	volume = {32},
 	pages = {1573--1375},
-	journal = {Stat Comp},
+	journal = {Statistics and Computing},
 	author = {Säilynoja, Teemu and Bürkner, Paul Christian and Vehtari, Aki},
   doi = {10.1007/s11222-022-10090-6},
 	year = {2022}
@@ -117,7 +117,7 @@ journal = {Journal of Open Source Software}
 title = {Detecting and diagnosing prior and likelihood sensitivity with power-scaling},
 author = {Noa Kallioinen and Topi Paananen and Paul Christian Bürkner and Aki Vehtari},
 year = {2023},
-journal = {Stat Comp},
+journal = {Statistics and Computing},
 volume = {34},
 issue = {57},
 doi = {10.1007/s11222-023-10366-5},
@@ -128,7 +128,7 @@ encoding = {UTF-8},
   author = {Paananen, T. and Piironen, J. and Bürkner, P. C. and Vehtari, A.},
   year = 2021,
   title = {Implicitly adaptive importance sampling},
-  journal = {Stat Comp},
+  journal = {Statistics and Computing},
   volume = 31,
   number = 16,
   doi= {10.1007/s11222-020-09982-2},
@@ -137,7 +137,7 @@ encoding = {UTF-8},
 @article{Gelman_2019,
 author = {Andrew Gelman and Ben Goodrich and Jonah Gabry and Aki Vehtari},
 title = {R-squared for {Bayesian} regression models},
-journal = {Am Stat},
+journal = {The American Statistician},
 doi={10.1080/00031305.2018.1549100},
 volume = {73},
 number = {3},
@@ -297,7 +297,7 @@ year = {2023}
   title = {Visualization in Bayesian workflow},
   author = {Jonah Gabry and Daniel Simpson and Aki Vehtari and Michael Betancourt and Andrew Gelman},
   year = {2019},
-  journal = {J. R. Stat. Soc. A},
+  journal = {Journal of the Royal Statistical Society Series A},
   volume = {182},
   issue = {2},
   pages = {389-402},
