### 2026-02-23 a1e0c1c
Ensure correct capitalization in titles
