+++ paper/references.bib
@@ -246,7 +246,7 @@ url = {https://plot.ly},
          number = {43},
          pages = {1864},
          author = {Daniel Foreman-Mackey and Will M. Farr and Manodeep Sinha and Anne M. Archibald and David W. Hogg and Jeremy S. Sanders and Joe Zuntz and Peter K. g. Williams and Andrew R. j. Nelson and Miguel de Val-Borro and Tobias Erhardt and Ilya Pashchenko and Oriol Abril Pla},
-         title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC},
+         title = {{emcee} v3: A {Python} ensemble sampling toolkit for affine-invariant {MCMC}},
          journal = {Journal of Open Source Software} }
 
 
@@ -286,7 +286,7 @@ year = {2023}
 }
 
 @Misc{bayesplot0,
-  title = {bayesplot: Plotting for Bayesian Models},
+  title = {bayesplot: Plotting for {Bayesian} Models},
   author = {Jonah Gabry and Tristan Mahr},
   year = {2025},
   note = {R package version 1.15.0},
@@ -294,7 +294,7 @@ year = {2023}
 }
 
 @Article{bayesplot1,
-  title = {Visualization in Bayesian workflow},
+  title = {Visualization in {Bayesian} workflow},
   author = {Jonah Gabry and Daniel Simpson and Aki Vehtari and Michael Betancourt and Andrew Gelman},
   year = {2019},
   journal = {Journal of the Royal Statistical Society Series A},
