### 2026-02-23 126d520
Add missing proceedings book title data
