+++ paper/references.bib
@@ -342,6 +342,7 @@ publisher = {Association for Computing Machinery},
 address = {New York, NY, USA},
 url = {https://doi.org/10.1145/2858036.2858558},
 doi = {10.1145/2858036.2858558},
+booktitle = {Proceedings of the 2016 CHI Conference on Human Factors in Computing Systems},
 pages = {5092–5103},
 numpages = {12},
 keywords = {uncertainty visualization, transit predictions, mobile interfac-es, end-user visualization, dotplots},
