### 2026-02-23 ff082d5
Add missing DOIs for CRAN packages
