+++ paper/references.bib
@@ -291,6 +291,7 @@ year = {2023}
   year = {2025},
   note = {R package version 1.15.0},
   url = {https://mc-stan.org/bayesplot/},
+  doi = {10.32614/cran.package.bayesplot},
 }
 
 @Article{bayesplot1,
@@ -320,6 +321,7 @@ year = {2023}
   year = {2025},
   note = {R package version 2.9.0},
   url = {https://mc-stan.org/loo/},
+  doi = {10.32614/cran.package.loo},
 }
 
 @article{Padilla_2021,
