### 2026-02-23 #9889 @matt-graham
Hi @aloctavodia - I have opened a pull-request (arviz-devs/arviz#2537) with some requested final edits to the paper and references, including fixing the DOIs flagged as missing in the reference check results in https://github.com/openjournals/joss-reviews/issues/9889#issuecomment-3946491077 which I think are both valid DOIs for the references mentioned.
