### 2026-02-24 378f772
Edits to JOSS paper (#2537)
* Define acronyms on first usage

* Use ssentence case for section headers

* Correct verb conjugation

* Use em dash in place of hyphen

* Consistently use non-code formatted project name

* Expand abbreviated journal names

* Ensure correct capitalization in titles

* Add missing proceedings book title data

* Add missing DOIs for CRAN packages
