+++ paper/references.bib
@@ -58,7 +58,7 @@ journal = {Journal of Open Source Software}
                Paul Horsfall and
                Noah D. Goodman},
   title     = {Pyro: Deep Universal Probabilistic Programming},
-  journal   = {J. Mach. Learn. Res.},
+  journal   = {Journal of Machine Learning Research},
   volume    = {20},
   pages     = {28:1--28:6},
   year      = {2019},
@@ -86,7 +86,7 @@ journal = {Journal of Open Source Software}
 @article{Vehtari_2017,
   title={Practical {Bayesian} model evaluation using leave-one-out cross-validation and {WAIC}},
   author={Vehtari, Aki and Gelman, Andrew and Gabry, Jonah},
-  journal={Stat Comp},
+  journal={Statistics and Computing},
   doi={10.1007/s11222-016-9696-4},
   volume={27},
   pages={1413--1432},
@@ -107,7 +107,7 @@ journal = {Journal of Open Source Software}
 	title = {Graphical test for discrete uniformity and its applications in goodness-of-fit evaluation and multiple sample comparison},
 	volume = {32},
 	pages = {1573--1375},
-	journal = {Stat Comp},
+	journal = {Statistics and Computing},
 	author = {Säilynoja, Teemu and Bürkner, Paul Christian and Vehtari, Aki},
   doi = {10.1007/s11222-022-10090-6},
 	year = {2022}
@@ -117,7 +117,7 @@ journal = {Journal of Open Source Software}
 title = {Detecting and diagnosing prior and likelihood sensitivity with power-scaling},
 author = {Noa Kallioinen and Topi Paananen and Paul Christian Bürkner and Aki Vehtari},
 year = {2023},
-journal = {Stat Comp},
+journal = {Statistics and Computing},
 volume = {34},
 issue = {57},
 doi = {10.1007/s11222-023-10366-5},
@@ -128,7 +128,7 @@ encoding = {UTF-8},
   author = {Paananen, T. and Piironen, J. and Bürkner, P. C. and Vehtari, A.},
   year = 2021,
   title = {Implicitly adaptive importance sampling},
-  journal = {Stat Comp},
+  journal = {Statistics and Computing},
   volume = 31,
   number = 16,
   doi= {10.1007/s11222-020-09982-2},
@@ -137,7 +137,7 @@ encoding = {UTF-8},
 @article{Gelman_2019,
 author = {Andrew Gelman and Ben Goodrich and Jonah Gabry and Aki Vehtari},
 title = {R-squared for {Bayesian} regression models},
-journal = {Am Stat},
+journal = {The American Statistician},
 doi={10.1080/00031305.2018.1549100},
 volume = {73},
 number = {3},
@@ -246,7 +246,7 @@ url = {https://plot.ly},
          number = {43},
          pages = {1864},
          author = {Daniel Foreman-Mackey and Will M. Farr and Manodeep Sinha and Anne M. Archibald and David W. Hogg and Jeremy S. Sanders and Joe Zuntz and Peter K. g. Williams and Andrew R. j. Nelson and Miguel de Val-Borro and Tobias Erhardt and Ilya Pashchenko and Oriol Abril Pla},
-         title = {emcee v3: A Python ensemble sampling toolkit for affine-invariant MCMC},
+         title = {{emcee} v3: A {Python} ensemble sampling toolkit for affine-invariant {MCMC}},
          journal = {Journal of Open Source Software} }
 
 
@@ -286,18 +286,19 @@ year = {2023}
 }
 
 @Misc{bayesplot0,
-  title = {bayesplot: Plotting for Bayesian Models},
+  title = {bayesplot: Plotting for {Bayesian} Models},
[…]
