### 2026-02-24 #9889 @aloctavodia
Thanks, @matt-graham, and thanks to all the reviewers for your time and suggestions.
As this is a major milestone for ArviZ, the 1.0 release. We are going to take a couple of days before doing the release. I hope that's fine.
