### 2026-02-24 #9889 @matt-graham
> Thanks, [@matt-graham](https://github.com/matt-graham), and thanks to all the reviewers for your time and suggestions. As this is a major milestone for ArviZ, the 1.0 release. We are going to take a couple of days before doing the release. I hope that's fine.

Hi @aloctavodia - yes definitely, taking a few days before doing the release is fine! I've set a reminder to myself to check in on this at end of this week but if you need more time than this that is not a problem.
