### 2026-03-02 2e6807d
bump version to 1.0
