### 2026-03-02 bf63845
bump version to 1.0 (#2542)
