+++ src/arviz/__init__.py
@@ -60,7 +60,7 @@ except ModuleNotFoundError as err:
 info += _status
 
 # define version last so it isn't overwritten by the respective attribute in the imported libraries
-__version__ = "1.0.0rc0"
+__version__ = "1.0.0"
 
 info = f"Status information for ArviZ {__version__}\n\n{info}"
 
