### 2026-03-02 #9889 @OriolAbril
@matt-graham we have just released the [1.0 version on PyPI](https://pypi.org/project/arviz/1.0.0/).

* Version tag: [v1.0.0](https://github.com/arviz-devs/arviz/tree/v1.0.0)
* Zenodo entry with matching title and authors: [10.5281/zenodo.18837209](https://doi.org/10.5281/zenodo.18837209)
