### 2026-03-03 a4561ca
fix: handle race condition in daily warning write
