### 2026-03-03 e8c2707
add Apache license headers (#2534)
