+++ src/arviz/__init__.py
@@ -1,3 +1,16 @@
+# ﻿Copyright ArviZ contributors
+#
+# Licensed under the Apache License, Version 2.0 (the "License");
+# you may not use this file except in compliance with the License.
+# You may obtain a copy of the License at
+#
+#     http://www.apache.org/licenses/LICENSE-2.0
+#
+# Unless required by applicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS,
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+# See the License for the specific language governing permissions and
+# limitations under the License.
 # pylint: disable=unused-import,unused-wildcard-import,wildcard-import,invalid-name
 """Expose features from _ArviZverse_ refactored packages together in the ``arviz`` namespace."""
 
