### 2026-03-03 7efe039
style: fix encoding and trailing whitespace in __init__
