+++ arviz/utils.py
@@ -1,13 +1,11 @@
 # pylint: disable=too-many-nested-blocks
 """General utilities."""
-import datetime
 import functools
 import importlib
 import importlib.resources
 import re
 import warnings
 from functools import lru_cache
-from pathlib import Path  # Added this for Path()
 
 import matplotlib.pyplot as plt
 import numpy as np
