### 2026-03-04 f959096
Update arviz/__init__.py
Co-authored-by: Oriol Abril-Pla <oriol.abril.pla@gmail.com>
