### 2026-03-04 81f6bcb
chore: bump version to 0.23.4
